requirements.txt
requests>=2.31.0
pyyaml>=6.0
psutil>=5.9.0
schedule>=1.2.0
python-daemon>=3.0.1
lockfile>=0.12.2 